refactor: Move common and auth packages to global namespace
