+++ apps/backend/src/main/java/com/sos/backend/global/common/config/SecurityConfig.java
@@ -1,7 +1,7 @@
-package com.sos.backend.common.config;
+package com.sos.backend.global.common.config;
 
-import com.sos.backend.auth.JwtFilter;
-import com.sos.backend.auth.JwtProvider;
+import com.sos.backend.global.auth.JwtFilter;
+import com.sos.backend.global.auth.JwtProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
