+++ apps/backend/src/main/java/com/sos/backend/global/common/exception/CustomException.java
@@ -1,4 +1,4 @@
-package com.sos.backend.common.exception;
+package com.sos.backend.global.common.exception;
 
 import lombok.Getter;
